+++ src/pages/index/system/role.tsx
@@ -4,10 +4,12 @@ import { useAppMessage } from '@zealous-admin/layout/index'
 import {
   Button,
   Card,
+  Flex,
   Form,
   Input,
   Modal,
   Radio,
+  Space,
   Switch,
   Table,
 } from 'antd'
@@ -151,33 +153,21 @@ export default function SystemRole() {
 
   return (
     <div className="app-container">
-      <Card
-        className="filter-container"
-        title="筛选搜索"
-        extra={(
-          <>
-            <Button onClick={() => setListQuery({ ...listQuery, pageNum: 1, keyword: '' })} style={{ marginRight: 15 }}>重置</Button>
-            <Button type="primary" onClick={() => setListQuery({ ...listQuery, pageNum: 1 })}>查询搜索</Button>
-          </>
-        )}
-      >
-        <Form layout="inline">
-          <Form.Item label="输入搜索：">
+      <Card>
+        <Flex justify="space-between" align="center" style={{ marginBottom: 16 }}>
+          <Space>
             <Input
               value={listQuery.keyword}
               onChange={e => setListQuery({ ...listQuery, keyword: e.target.value })}
               placeholder="角色名称"
-              style={{ width: 200 }}
+              style={{ width: 220 }}
               allowClear
             />
-          </Form.Item>
-        </Form>
-      </Card>
-
-      <Card
-        title="数据列表"
-        extra={<Button type="primary" icon={<PlusOutlined />} onClick={handleAdd}>添加</Button>}
-      >
+            <Button type="primary" onClick={() => setListQuery({ ...listQuery, pageNum: 1 })}>查询</Button>
+            <Button onClick={() => setListQuery({ ...listQuery, pageNum: 1, keyword: '' })}>重置</Button>
+          </Space>
+          <Button type="primary" icon={<PlusOutlined />} onClick={handleAdd}>添加</Button>
+        </Flex>
         <Table
           columns={columns}
           dataSource={list}